for balance

+++ Assets/Scripts/Zenject/MonoInstallers/ShopSystemInstaller.cs
@@ -7,8 +7,8 @@ public class ShopSystemInstaller : MonoInstaller
 {
     [Header("Upgrades")]
     public List<int> DamageUpgrades;
-    public List<int> SpeedUpgrades;
-    public List<int> CastUpgrades;
+    public List<float> SpeedUpgrades;
+    public List<float> CastUpgrades;
     [Header("Prices")]
     public List<int> DamageUpgradesPrices;
     public List<int> SpeedUpgradesPrices;
@@ -17,8 +17,8 @@ public class ShopSystemInstaller : MonoInstaller
     public override void InstallBindings()
     {
         Container.Bind<List<int>>().WithId("DamageUpgrades").FromInstance(DamageUpgrades).AsCached();
-        Container.Bind<List<int>>().WithId("SpeedUpgrades").FromInstance(SpeedUpgrades).AsCached();
-        Container.Bind<List<int>>().WithId("CastUpgrades").FromInstance(CastUpgrades).AsCached();
+        Container.Bind<List<float>>().WithId("SpeedUpgrades").FromInstance(SpeedUpgrades).AsCached();
+        Container.Bind<List<float>>().WithId("CastUpgrades").FromInstance(CastUpgrades).AsCached();
         
         Container.Bind<List<int>>().WithId("DamageUpgradesPrices").FromInstance(DamageUpgradesPrices).AsCached();
         Container.Bind<List<int>>().WithId("SpeedUpgradesPrices").FromInstance(SpeedUpgradesPrices).AsCached();
